# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DrépaConnect API — plateforme de lutte contre la drépanocytose
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List medications for a patient
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Medication list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: createMedication
      tags: [medications]
      summary: Add medication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationInput"
      responses:
        "201":
          description: Medication added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /medications/{id}:
    patch:
      operationId: updateMedication
      tags: [medications]
      summary: Update medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Delete medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /medications/{id}/take:
    post:
      operationId: takeMedication
      tags: [medications]
      summary: Mark medication as taken
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"

  # SCREENING CENTERS
...
  /medical-records:
    get:
      operationId: listMedicalRecords
      tags: [medical-records]
      summary: List medical records
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Records list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MedicalRecord"
    post:
      operationId: createMedicalRecord
      tags: [medical-records]
      summary: Create medical record entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicalRecordInput"
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"
  /medical-records/{id}:
    get:
      operationId: getMedicalRecord
      tags: [medical-records]
      summary: Get medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medical record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalRecord"

  # BLOOD DONORS
  /donors:
    get:
      operationId: listDonors
      tags: [donors]
      summary: List available blood donors
      parameters:
        - name: bloodType
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Donors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodDonor"
    post:
      operationId: registerDonor
      tags: [donors]
      summary: Register as blood donor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodDonorInput"
      responses:
        "201":
          description: Donor registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodDonor"
  /donors/{id}/availability:
    patch:
      operationId: updateDonorAvailability
      tags: [donors]
      summary: Toggle donor availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorAvailabilityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodDonor"

  # BLOOD REQUESTS
...
  /education:
    get:
      operationId: listEducationModules
      tags: [education]
      summary: List educational modules
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: targetAudience
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Education modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EducationModule"
  /education/{id}:
    get:
      operationId: getEducationModule
      tags: [education]
      summary: Get education module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EducationModule"

  # STATISTICS
...
components:
  schemas:
...
    Medication:
      type: object
      required: [id, patientId, name, dosage, frequency, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        times:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        active:
          type: boolean
        lastTaken:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    MedicationInput:
      type: object
      required: [patientId, name, dosage, frequency]
      properties:
        patientId:
          type: integer
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        times:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    MedicationUpdate:
      type: object
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        times:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        active:
          type: boolean
        notes:
          type: ["string", "null"]

    # SCREENING CENTER
...
    MedicalRecord:
      type: object
      required: [id, patientId, type, title, date, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        type:
          type: string
          enum: [CONSULTATION, LAB_RESULT, TRANSFUSION, VACCINATION, HOSPITALIZATION, OTHER]
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        doctorName:
          type: ["string", "null"]
        facility:
          type: ["string", "null"]
        results:
          type: ["string", "null"]
        hemoglobinLevel:
          type: ["number", "null"]
        createdAt:
          type: string
    MedicalRecordInput:
      type: object
      required: [patientId, type, title, date]
      properties:
        patientId:
          type: integer
        type:
          type: string
          enum: [CONSULTATION, LAB_RESULT, TRANSFUSION, VACCINATION, HOSPITALIZATION, OTHER]
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        doctorName:
          type: ["string", "null"]
        facility:
          type: ["string", "null"]
        results:
          type: ["string", "null"]
        hemoglobinLevel:
          type: ["number", "null"]

    # BLOOD DONOR
    BloodDonor:
      type: object
      required: [id, userId, bloodType, city, available, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: ["string", "null"]
        bloodType:
          type: string
          enum: [A_POS, A_NEG, B_POS, B_NEG, AB_POS, AB_NEG, O_POS, O_NEG]
        city:
          type: string
        province:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        available:
          type: boolean
        lastDonation:
          type: ["string", "null"]
        donationCount:
          type: integer
        createdAt:
          type: string
    BloodDonorInput:
      type: object
      required: [bloodType, city]
      properties:
        bloodType:
          type: string
          enum: [A_POS, A_NEG, B_POS, B_NEG, AB_POS, AB_NEG, O_POS, O_NEG]
        city:
          type: string
        province:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
    DonorAvailabilityUpdate:
      type: object
      required: [available]
      properties:
        available:
          type: boolean

    # BLOOD REQUEST
...
    EducationModule:
      type: object
      required: [id, title, category, targetAudience, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [BASICS, GENETICS, SYMPTOMS, TREATMENT, NUTRITION, PREVENTION, SCHOOL, FAMILY]
        targetAudience:
          type: string
          enum: [ALL, CHILDREN, ADULTS, STUDENTS, TEACHERS, PARENTS, DOCTORS]
        summary:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        readTime:
          type: ["integer", "null"]
        createdAt:
          type: string

    # STATISTICS
...